Move reason code checks
